Pre merge code reorg, cleanup, small bug fix in centroid count overflow
