+++ FastCG/generators/__init__.py
@@ -2,3 +2,4 @@
 # Path: __init__.py
 from FastCG.generators.GeneratorBase import GeneratorBase
 from FastCG.generators.ProbKnnGenerator import ProbKnnGenerator
+from FastCG.generators.Optika import Optika
