+++ FastCG/utils/__init__.py
@@ -2,3 +2,6 @@
 # Path: __init__.py
 import FastCG.utils.util_functions as util_functions
 from FastCG.utils.SmartCondition import SmartCondition
+from FastCG.utils.Benchmark import Benchmark
+from FastCG.utils.Face import Face
+from FastCG.utils.GS import GrowingSpheres as GS
